refactor(trackersSlice): refactor trackersSlice reducers tests
